VRS 2013.1.0.8
Bug
[ACL-193] - En la ventana de usuario no va la paginación.
[ACL-194] - En las unidades de tratamientos no deja meter decimales.
Improvement
[ACL-196] - Que la edad se calcule al terminar de poner la fecha de nacimiento.
Task
[ACL-186] - Hacer que en las búsquedas de profesional solo aparezcan los activos
[ACL-189] - Incluir un indicador de "inactivo" en el professional
[ACL-192] - Crear un sistema de plantillas

+++ AriClinic/AriCliWeb/TreatmentForm.aspx.cs
@@ -204,7 +204,7 @@ public partial class TreatmentForm : System.Web.UI.Page
         rdpTreatmentDate.SelectedDate = tr.TreatmentDate;
         txtRecommend.Text = tr.Recommend;
         if (tr.Quantity != 0)
-            txtQuantity.Value = tr.Quantity;
+            txtQuantity.Value = (double)tr.Quantity;
     }
 
     protected void UnloadData(Treatment tr)
@@ -215,7 +215,7 @@ public partial class TreatmentForm : System.Web.UI.Page
         if (visit != null)
             tr.BaseVisit = visit;
         tr.Recommend = txtRecommend.Text;
-        if (txtQuantity.Text != "") tr.Quantity = (int)txtQuantity.Value;
+        if (txtQuantity.Text != "") tr.Quantity = (decimal)txtQuantity.Value;
         if (rdcProfessional.SelectedValue != "")
         {
             tr.Professional = CntAriCli.GetProfessional(int.Parse(rdcProfessional.SelectedValue), ctx);
